refactor(backend): restructure monolithic files into modular architecture with automated bundling  (#16)
### Modular Architecture
- **API Layer**: Split `idev-suitetools-api.ts` into modular components
  - Extracted `SuiteToolsApi`, `SuiteToolsApiGet`, `SuiteToolsApiPost`, etc. into `/api/` directory
  - Maintained clean entry point with proper error handling using shared error classes

- **App Layer**: Refactored `idev-suitetools-app.ts` into `/app/` directory structure
  - Separated `SuiteToolsApp` and `SuiteToolsAppView` classes into dedicated modules
  - Maintained entry point simplicity with proper routing logic

- **Common Layer**: Decomposed `idev-suitetools-common.ts` into `/common/` structure
  - Split into focused modules: `SuiteToolsCommon`, `SuiteToolsCommonSettings`, `SuiteToolsCommonJobs`
  - Created `/common/library/` with specialized NetSuite operation classes:
    - `SuiteToolsCommonLibraryNetSuiteFile` - File operations
    - `SuiteToolsCommonLibraryNetSuiteHttp` - HTTP requests
    - `SuiteToolsCommonLibraryNetSuiteRecord` - Record CRUD operations
    - `SuiteToolsCommonLibraryNetSuiteScript` - Script management
    - `SuiteToolsCommonLibraryNetSuiteSearch` - Search functionality
    - `SuiteToolsCommonLibraryNetSuiteSuiteQl` - SuiteQL queries
    - `SuiteToolsCommonLibraryNetSuiteTask` - Task management

### Automated Build System
- **Deployment Bundling**: Created `build-deployments.mjs` with esbuild integration
  - Bundles modular TypeScript into NetSuite-compatible single files
  - Converts CommonJS output to AMD format required by NetSuite
  - Handles N/ module reference mapping and dependency resolution
  - Generates deployment-ready files: `idev-suitetools-api.js` (425KB) and `idev-suitetools-app.js` (49KB)

### Benefits
- **Maintainability**: Clear separation of concerns with focused, testable modules
- **Developer Experience**: Easier navigation and modification of specific functionality
- **Build Automation**: Single command creates deployment-ready NetSuite bundles

### Technical Details
- Build system handles NetSuite-specific AMD conversion and N/ module mapping
