+++ backend/package.json
@@ -18,7 +18,9 @@
   "license": "GPL-3.0-or-later",
   "scripts": {
     "deploy": "suitecloud project:deploy",
-    "build": "yarn tsc",
+    "build": "yarn build:tsc && yarn build:deployments",
+    "build:tsc": "tsc",
+    "build:deployments": "node build-deployments.mjs",
     "validate": "suitecloud project:validate",
     "watch": "tsc --watch",
     "lint": "eslint . --report-unused-disable-directives --max-warnings 0",
@@ -36,6 +38,7 @@
     "@types/node": "^22.13.0",
     "@typescript-eslint/eslint-plugin": "^8.22.0",
     "@typescript-eslint/parser": "^8.22.0",
+    "esbuild": "*",
     "eslint": "*",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-suitescript": "^1.3.2",
